+++ core/src/main/scala/org/opentorah/metadata/HasName.scala
@@ -1,7 +1,6 @@
 package org.opentorah.metadata
 
 import org.opentorah.util.{ClassName, Effects}
-import org.opentorah.xml.{ElementsTo, From, Parser}
 
 trait HasName(nameOverride: Option[String]):
   final def name: String = nameOverride.getOrElse(defaultName)
@@ -18,21 +17,11 @@ object HasName:
     self: HasName =>
     final override protected def defaultName: String = ClassName.get(this)
 
-  def load[K <: HasName, M](
-    from: From,
-    content: ElementsTo[M],
-    keys: Seq[K],
-    hasName: (M, String) => Boolean
-  ): Parser[Map[K, M]] = for
-    metadatas: Seq[M] <- load[M](from, content)
-    result: Map[K, M] <- mapByName(keys, metadatas, hasName)
-  yield result
-
   def mapByName[K <: HasName, M](
     keys: Seq[K],
     metadatas: Seq[M],
     hasName: (M, String) => Boolean
-  ): Parser[Map[K, M]] = for
+  ): Effects.IO[Map[K, M]] = for
     result: Seq[(K, M)] <- Effects.collectAll(metadatas.map(metadata =>
       find(
         keys,
@@ -43,16 +32,11 @@ object HasName:
     _ <- checkNoUnmatchedKeys(keys.toSet -- result.map(_._1).toSet)
   yield result.toMap
 
-  def load[M](
-    from: From,
-    content: ElementsTo[M]
-  ): Parser[Seq[M]] = content.wrappedSeq(from.name).parse(from)
-
   def bind[K, M](
     keys: Seq[K],
     metadatas: Seq[M],
     getKey: M => K
-  ): Parser[Map[K, M]] =
+  ): Effects.IO[Map[K, M]] =
     val result: Map[K, M] = metadatas.map(metadata => getKey(metadata) -> metadata).toMap
     for
       _ <- checkNoUnmatchedKeys(keys.toSet -- result.keySet)
@@ -61,7 +45,7 @@ object HasName:
   def find[K <: HasName](
     keys: Seq[K],
     names: Names
-  ): Parser[K] = find(
+  ): Effects.IO[K] = find(
     keys = keys,
     metadata = names,
     hasName = (names: Names, name: String) => names.hasName(name)
@@ -77,7 +61,7 @@ object HasName:
     keys: Seq[K],
     metadata: M,
     hasName: (M, String) => Boolean
-  ): Parser[K] =
+  ): Effects.IO[K] =
     val result: Seq[K] = keys.filter(key => hasName(metadata, key.name))
     for
       _ <- Effects.check(result.nonEmpty, s"Unmatched metadata: $metadata")