+++ core/src/main/scala/org/opentorah/xml/Elements.scala
@@ -1,3 +1,0 @@
-package org.opentorah.xml
-
-type Elements = Seq[Element]